Fix edge case due to floating point imprecision
It was possible for the intersection point be coincident to the bisector
origin, where the origin was consider to be to the right of the
bisector. In this case, the intersection point would fall outside of the
domain of Cqs-, missing the intersection.

This was due to the way the comparison worked, where it was consider to
be right of Cqs+ if the point was below. This was replaced by
calculating the intersection with a horizontal line instead.

Still not sure if this solution actually always work.

Also a couple of random changes, like using `Point::new` everywere, and
making `circumcenter` return a option.